move calcquality tests
